requirements.txt
# Enterprise RAG Requirements - OpenAI + Chroma with flexible data sources
# Core LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Fixed providers - no switching needed
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Data processing
pandas>=2.0.0
python-dotenv>=1.0.0

# Optional: S3 support
boto3>=1.34.0

# Optional: Confluence support
atlassian-python-api>=3.41.0
beautifulsoup4>=4.12.0 